requirements.txt
# Core Dependencies
pandas>=1.4.0
numpy>=1.21.0
scikit-learn>=1.0.0
plotly>=5.5.0
dash>=2.0.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
pydantic>=1.9.0
email-validator>=1.1.3
python-dotenv>=0.19.0
aiofiles>=0.8.0

# Data Dependencies
yfinance>=0.1.70
ccxt>=2.0.0
python-binance>=1.0.0
ibapi>=9.81.1
websockets>=10.0
aiohttp>=3.8.0

# Monitoring
psutil>=5.8.0
prometheus-client>=0.12.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.16.0
pytest-cov>=2.12.0
pytest-mock>=3.6.0
httpx>=0.24.0
aioresponses>=0.7.0
hypothesis>=6.75.3
freezegun>=1.2.0
pytest-timeout>=2.1.0
pytest-recording>=0.12.0

# Documentation
sphinx>=4.4.0
sphinx-rtd-theme>=1.0.0
sphinx-autodoc-typehints>=1.12.0

# Development Tools
black>=22.0.0
isort>=5.10.0
mypy>=0.910
pylint>=2.12.0
flake8>=4.0.0
autopep8>=1.6.0

# Security
bandit>=1.7.0
safety>=1.10.3
cryptography>=36.0.0

# Data Processing
numpy>=1.21.0
pandas>=1.4.0
scipy>=1.7.0
statsmodels>=0.13.0
scikit-learn>=1.0.0
ta>=0.7.0  # Technical Analysis

# Web Interface
fastapi>=0.68.0
uvicorn>=0.15.0
jinja2>=3.0.0
aiofiles>=0.8.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyjwt>=2.3.0

# Utilities
python-dateutil>=2.8.2
pytz>=2021.3
requests>=2.26.0
tqdm>=4.62.0
tenacity>=8.0.0
PyYAML>=6.0.0
apscheduler>=3.9.0
celery>=5.2.0
redis>=4.1.0
gspread>=5.4.0
oauth2client>=4.1.3

# Database
sqlalchemy>=1.4.0
alembic>=1.7.0
psycopg2-binary>=2.9.0
asyncpg>=0.25.0

# Optimization
numba>=0.54.0
cython>=0.29.0
pyarrow>=6.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.5.0
bokeh>=2.4.0